requirements.txt
fastapi  
pydantic  
langchain-core  
langchain-openai  
requests  
python-dotenv  
opencage-geocoder  
sqlite3  
uvicorn